Singleplayer mode
The gamemode now functions properly when there is only 1 player on the
server. That player will spawn as Runner and will not take damage,
allowing them to explore the map and practice their runs. Added a !r
command to respawn the player if he gets stuck.

Players can download the gamemode and start it up on a listen server (not
sure if singleplayer works because of net messages and whatnot). Players
can join empty servers and practice their runs, so there's something to do
when waiting for friends to join.