Added thirdperson
Added a thirdperson convar for clients to play the game in thirdperson.
Fleshed it out by adding X Y Z offset convars to do over-the-shoulder
stuff or just customise the view in general. Can be changed through
settings menu.

Bound F1 to Help and F2 to Settings.

Modified weapon_cs_base so that it supported thirdperson mode.