Update init.lua
Changed GM:CanPlayerSuicide to prevent people from death dodging, as this will disable their ability to suicide as death. Still disables suiciding as spec as well, but calls it in a different way which doesn't matter. I added the if ROUND:GetTimer() using your hook, but for some reason it does not work and the player can suicide. You can uncomment it and mess with it if you can figure it out.