Fix nomination cooldown
Previously the nomination cooldown didn't work. ply.LastNom was set before it was checked, so players would have to wait one second after trying to nominate for the first time. In addition, ply.LastNom was only set once, so after the initial one second passed players are able to spam it as much as they like.

This PR changes it so the cooldown only takes effect after the first nomination, and it correctly sets ply.LastNom each subsequent nomination so players have to wait 1 second between nominations.