fixed a bug where players entering a area from X+ or Y+ wouldn't be detected as long as ply:GetPos() isn't inside
this should be some logic misatake that left these cases out of
CuboidOverlap(), and when ply:GetPos() hits the area PlayerInCuboid()
returns true, and ZONE:Tick() carries on as normal.
I found this when I was creating the Barrier zone, when some bugs
occurred that weren't happening becouse of my code.