improved border for zone detection
The sphere creathed by the other method was not good for thin zones
where the sphere whould extend far outside of the zone on the center of
a face and get just outside of it on the edges.

This caused a bug with the barrier zone where the server would not have
time to call the hook "DeathrunPlayerExitedZone"  before the player left
the circle, which prevented the hook from being called at all. This
happened when the player entered the zone through a edge, where the
sphere is close to the actual zone.
I could have just enlarged the sphere a bit, but this seemed like a
better fix...